Replaces "pod update" with "pod repo update"
"pod update" checks if there are new versions of the installed pods available and uses them instead.

"pod repo update" updates the spec repositories to ensure the pods specified in the lockfile are found by the CocoaPods.

Using "pod update" is potentially dangerous as it might change the version of the pods in use.